requirements.txt
absl-py==0.9.0
aenum==2.2.3
argcomplete==1.9.5
asn1crypto==0.24.0
astor==0.8.1
attrs==19.1.0
autopep8==1.5
backcall==0.1.0
Beaker==1.10.0
beautifulsoup4==4.8.2
bitstruct==8.9.0
bleach==3.1.0
blivet==3.1.4
blivet-gui==2.1.11
Brlapi==0.6.7
cachetools==4.1.0
cantools==33.1.1
cassandra-driver==3.18.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
click-conf==0.0.1
click-stream==0.0.10
conda==4.6.14
configparser==3.7.1
coverage==4.5.1
cryptography==2.6.1
cson==0.7
cupshelpers==1.0
curlew==0.2.4
cycler==0.10.0
cytoolz==0.9.0.1
dbfread==2.0.7
dbus-python==1.2.8
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
discrete==0.0.2
diskcache==4.1.0
distro==1.4.0
elasticsearch==7.5.1
elasticsearch-loader==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
factorial==1.0.0
Flask==1.0.3
flask-ldap3-login==0.9.16
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
fros==1.1
frozendict==1.2
future==0.18.2
gast==0.2.2
geocoder==1.38.1
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpg==1.12.0
grpcio==1.28.1
h5py==2.10.0
humanize==0.5.1
idna==2.7
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.1.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jellyfish==0.7.2
Jinja2==2.10.1
joblib==0.14.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==0.35.6
jupyterlab-server==0.2.0
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langtable==0.0.43
ldap3==2.6
lesscpy==0.14.0
libcomps==0.1.14
llvmlite==0.31.0
lxml==4.5.0
Mako==1.1.0.dev0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.0
meson==0.52.0
mistune==0.8.4
mizani==0.5.4
mod-wsgi==4.6.4
nbconvert==5.5.0
nbformat==4.4.0
neo4j-driver==1.7.4
neo4jrestclient==2.1.1
neobolt==1.7.13
neotime==1.7.4
notebook==5.7.8
ntplib==0.3.3
numba==0.48.0
numpy==1.17.0
oauthlib==3.1.0
olefile==0.46
openpyxl==2.6.3
opt-einsum==3.2.0
ordered-set==2.0.2
palettable==3.1.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.4.0
Paste==2.0.3
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pid==2.2.3
Pillow==5.4.1
pipenv==2018.11.26
Platypus==1.0
plotnine==0.5.1
ply==3.11
productmd==1.23
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.11.3
ptyprocess==0.6.0
pwquality==1.4.0
py-cpuinfo==4.0.0
pyaml==19.12.0
pyarrow==0.16.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.18.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.74
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
Pygments==2.4.2
PyGObject==3.32.2
pyinotify==0.9.6
pykickstart==3.20
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyparted==3.11.2
PyQt5==5.13.1
PyQt5-sip==4.19.19
pyrsistent==0.15.2
pyshp==2.1.0
PySocks==1.6.8
python-augeas==0.5.0
python-can==3.3.2
python-dateutil==2.8.0
python-dotenv==0.13.0
python-meh==0.47
python-telegram-bot==12.7
pytz==2019.2
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3
pyzmq==18.0.1
qtconsole==4.5.1
ratelim==0.1.6
redis==3.2.1
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
rpm==4.14.2.1
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
schedule==0.6.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
selinux==2.9
Send2Trash==1.5.0
sepolicy==1.1
setools==4.1.1
setroubleshoot==1.1
simpleline==1.4
six==1.12.0
sklearn==0.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.8
soupsieve==2.0
speg==0.3
SQLAlchemy==1.3.4
stats==0.1.2a0
statsmodels==0.9.0
systemd-python==234
tabula-py==2.1.0
Tempita==0.5.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textparser==0.23.0
thrift==0.13.0
toolz==0.9.0
tornado==6.0.2
tqdm==4.37.0
traitlets==4.3.2
urllib3==1.25.7
vboxapi==1.0
virtualenv==16.0.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
